extract types and refactor
